Migrate ffapi-update-nodes to ffapi-update
- Export config settings to external file
- Remove OLSR related functions
- Add using template file instead of just overwrite old api-file.
  (Old functionality can be achieved using same file as template and destination file)
